Install by cloning GitHub repo

+++ R/uninstall_tmhmm.R
@@ -13,16 +13,17 @@ uninstall_pureseqtm <- function(
     stop("Cannot uninstall absent PureseqTM at folder '", folder_name, "'")
   }
   # Remove extracted folder
-  pureseqtm_folder <- file.path(folder_name, "PureseqTM_Package-master")
+  pureseqtm_folder <- file.path(folder_name, "PureseqTM_Package")
   if (dir.exists(pureseqtm_folder)) {
     unlink(pureseqtm_folder, recursive = TRUE, force = TRUE)
   }
   testthat::expect_false(dir.exists(pureseqtm_folder))
-
-  # Download zip if needed
-  zip_filename_path <- file.path(folder_name, "PureseqTM_Package.zip")
-  if (file.exists(zip_filename_path)) {
-    file.remove(zip_filename_path)
+  if (1 == 2) {
+    # Download zip if needed
+    zip_filename_path <- file.path(folder_name, "PureseqTM_Package.zip")
+    if (file.exists(zip_filename_path)) {
+      file.remove(zip_filename_path)
+    }
+    testthat::expect_false(file.exists(zip_filename_path))
   }
-  testthat::expect_false(file.exists(zip_filename_path))
 }
